fixed all warnings and new file structure
