+++ react-app/src/App.js
@@ -9,7 +9,6 @@ import LandingPage from './components/LandingPage';
 import ErrorPage from './components/ErrorPage';
 import { authenticate } from './store/session';
 import * as serverActions from "./store/server";
-import MessageComponent from './components/MessageComponent/MessageComponent'
 
 
 
@@ -27,7 +26,7 @@ function App() {
       })
       setLoaded(true);
     })();
-  }, [dispatch]);
+  }, [dispatch, user]);
 
   if (!loaded) {
     return null;
