+++ react-app/src/components/LandingPage/LandingPage.css
@@ -92,7 +92,7 @@
 }
 
 .top-row-div {
-    background-image: url();
+    /* background-image: url(); */
     border-width: 0;
     box-sizing: border-box;
     color: #000;
