+++ react-app/src/components/ServerList/ServersList.jsx
@@ -1,23 +1,35 @@
-import React from "react";
-import CreateServerModal from "../../components/CreateServer/CreateServer";
-import { useSelector } from "react-redux";
+import React, { useEffect, useState } from "react";
+import * as serverActions from "../../store/server";
+import CreateServerModal from "../../components/CreateServer/index";
+import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import './ServerList.css'
 
 const ServerList = () => {
     const userServers = useSelector(state => Object.values(state.server.userServers))
+    const dispatch = useDispatch();
+    const [isLoaded, setIsLoaded] = useState(false)
+    const user = useSelector(state => state.session.user);
 
-    return (
-        <div className="sb">
-            <div className="sb-icon-wrapper">
+    useEffect(() => {
+        const loaded = async () => {
+            await dispatch(serverActions.loadUsersServers(user.id))
+            setIsLoaded(true)
+        }
+        loaded()
+    }, [dispatch, user.id]);
+
+    return isLoaded && (
+        <div className="sl">
+            <div className="sl-icon-wrapper">
             {userServers.slice(0, 1).map(server => (
                     <Link key={server.id}  to={`/@me/${server.id}`}>
-                        <img src={server.icon} alt="" className="sb-server-icon" />
+                        <img src={server.icon} alt="" className="sl-server-icon" />
                     </Link>
                 ))}
                 {userServers.slice(1).map(server => (
                     <Link key={server.id} to={`/servers/${server.id}`}>
-                        <img src={server.icon} alt="" className="sb-server-icon" />
+                        <img src={server.icon} alt="" className="sl-server-icon" />
                     </Link>
                 ))}
             </div>
