+++ react-app/src/components/ServerNav/ServerNav.css
@@ -1,9 +1,4 @@
-.server-nav{
-    background-color: rgb(32,34,37);
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+.sn{
+    height: 100vh;
     width: 72px;
-    overflow-y: hidden;
 }
